refactor: 폴더별 constant파일 생성 및 분리, 중복코드 제거
