+++ src/features/admin/theme/api/theme-api.ts
@@ -1,7 +1,7 @@
 import { createApi, fakeBaseQuery } from '@reduxjs/toolkit/query/react';
 
 import type { SystemTheme } from '../types/theme-types';
-import { DEFAULT_SYSTEM_THEME } from '../../../../shared/constants';
+import { DEFAULT_SYSTEM_THEME } from '../constants/default-theme';
 
 interface UserThemePreference {
   mode: 'light' | 'dark';
