+++ src/features/auth/constants/auth-constants.ts
@@ -1,0 +1,4 @@
+export const AUTH_ENDPOINTS = {
+  REFRESH_TOKEN: '/auth/token/refresh',
+  LOGOUT: '/auth/logout',
+} as const;
